added all stuff with matching .meta files
included here is the synth with a lot of it working at the moment. Some ranges are not as they should be yet, but that will come in another commit. included is also the tree spawner script, currently being hooked up to a slider. the render pipeline is the one implemented by Julius.